+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Balanced Dashboard Changelog
 
+### master
+
+* Updated search results view (#1489)
+* Added keyboard shortcuts to open and close the search modal
+
 ### 1.1.1
 
 * Bug fixes for order views